Working towards the full standalone fortran code

+++ dynamics/sam++/fortran/diagnose.F90
@@ -13,7 +13,7 @@ contains
     implicit none
     integer(crm_iknd), intent(in) :: ncrms
     integer(crm_iknd) i,j,k,kb,kc,icrm
-    real(8) coef, coef1
+    real(r8) coef, coef1
     real(crm_rknd) tmp_lwp, tmp
 
     coef = 1./real(nx*ny,crm_rknd)
