+++ dynamics/sam++/fortran/press_rhs.F90
@@ -11,7 +11,7 @@ contains
     use params, only: dowallx, dowally
     implicit none
     integer(crm_iknd), intent(in) :: ncrms
-    real *8 dta,rdx,rdy,rdz,btat,ctat,rup,rdn
+    real(r8) dta,rdx,rdy,rdz,btat,ctat,rup,rdn
     integer(crm_iknd) i,j,k,ic,jc,kc, icrm
 
     if(dowallx.and.mod(rank,nsubdomains_x).eq.0) then
